+++ sieve.h
@@ -3,8 +3,8 @@
 
 #include <stdint.h>
 
-void sieveOpenMP(uint32_t n, int nThreads);
-void seqSieveOptimized(uint32_t n);
-void seqSieve(uint32_t n);
+void sieveOpenMP(uint64_t n, int nThreads);
+void seqSieveOptimized(uint64_t n);
+void seqSieve(uint64_t n);
 
 #endif
